Backend

+++ src/main/java/org/springframework/samples/petclinic/web/DepartmentController.java
@@ -2,15 +2,18 @@ package org.springframework.samples.petclinic.web;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.samples.petclinic.model.Department;
+import org.springframework.samples.petclinic.model.Role;
 import org.springframework.samples.petclinic.model.Team;
+import org.springframework.samples.petclinic.model.UserTW;
 import org.springframework.samples.petclinic.service.DepartmentService;
 import org.springframework.samples.petclinic.service.TeamService;
+import org.springframework.samples.petclinic.service.UserTWService;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,54 +27,70 @@ import org.springframework.web.bind.annotation.RestController;
 public class DepartmentController {
 	private final DepartmentService departmentService;
 	private final TeamService teamService;
+	private final UserTWService userTWService;
+
 	@Autowired
-	public DepartmentController(DepartmentService departmentService,TeamService teamService) {
+	public DepartmentController(DepartmentService departmentService, TeamService teamService,
+			UserTWService userTWService) {
 		this.departmentService = departmentService;
-		this.teamService=teamService;
+		this.teamService = teamService;
+		this.userTWService = userTWService;
 	}
-	
+
 	@InitBinder
 	public void setAllowedFields(WebDataBinder dataBinder) {
 		dataBinder.setDisallowedFields("id");
 	}
-	
-	@GetMapping(value = "/departments")
-	public List<Department> getDeparments(@RequestParam(required = false) String name){
-		List<Department> l=new ArrayList<>();
-		if(name==null) {
-			l = departmentService.getAllDepartments().stream().collect(Collectors.toList());
-            return l;
-		}
-		else {
-			l=departmentService.findDepartmentByName(name).stream().collect(Collectors.toList());
-			return l;
-		}
+
+	@GetMapping(value = "/api/departments")
+
+	public List<Department> getDeparments(HttpServletRequest r) {
+		List<Department> l = new ArrayList<>();
+		Integer teamId = (Integer) r.getSession().getAttribute("teamId");
+		l = teamService.findTeamById(teamId).getDepartments();
+		return l;
 	}
-	
-	@PostMapping(value = "/departments")
-	public ResponseEntity<String> postDeparments(@RequestParam(required = true) Integer teamId,@RequestBody Department department) {
+
+	@PostMapping(value = "/api/departments")
+	public ResponseEntity<String> postDeparments(@RequestParam(required = true) Integer teamId,
+			@RequestBody Department department, HttpServletRequest r) {
+
 		try {
-			Team team=teamService.findTeamById(teamId);
-			department.setTeam(team);
-			departmentService.saveDepartment(department);
-			return ResponseEntity.ok("Department create");
-		}
-		catch(DataAccessException d) {
+			Integer userId = (Integer) r.getSession().getAttribute("userId");
+			UserTW user = userTWService.findUserById(userId);
+			if (user.getRole().equals(Role.team_owner)) {
+
+				Team team = teamService.findTeamById(teamId);
+				department.setTeam(team);
+				departmentService.saveDepartment(department);
+				return ResponseEntity.ok("Department create");
+			} else {
+				return ResponseEntity.status(403).build();
+			}
+
+		} catch (DataAccessException d) {
 			return ResponseEntity.badRequest().build();
 		}
-	    
+
 	}
-	
-	@DeleteMapping(value = "/departments")
-	public ResponseEntity<String> deleteDeparments(@RequestParam(required = true) Integer departmentId) {
-		System.out.println(departmentId);
+
+	@DeleteMapping(value = "/api/departments")
+	public ResponseEntity<String> deleteDeparments(@RequestParam(required = true) Integer departmentId,
+			HttpServletRequest r) {
+
 		try {
-			departmentService.deleteDepartmentById(departmentId);
-	        return ResponseEntity.ok("Department delete");
-		}
-		catch(DataAccessException d) {
+			Integer userId = (Integer) r.getSession().getAttribute("userId");
+			UserTW user = userTWService.findUserById(userId);
+			if (user.getRole().equals(Role.team_owner)) {
+				departmentService.deleteDepartmentById(departmentId);
+				return ResponseEntity.ok("Department delete");
+			} else {
+				return ResponseEntity.status(403).build();
+			}
+
+		} catch (DataAccessException d) {
 			return ResponseEntity.notFound().build();
 		}
-	    
+
 	}
 }
